fix: rename cancell_my_reservation -> cancel_my_reservation and OrganizerEventUpdateView naming
