+++ backend/reservations/urls.py
@@ -4,7 +4,7 @@ from reservations.views.reservation_list import ParticipantsListView
 from reservations.views.reservation_status_update import ReservationStatusUpdateView
 from reservations.views.my_reservation import MyReservationsView
 from reservations.views.organizer_reservation_list import OrganizerReservationView
-from reservations.views.cancell_my_reservation import CancelMyReservation
+from reservations.views.cancel_my_reservation import CancelMyReservation
 from reservations.views.my_single_reservation_detail import MySingleReservationDetailView
 from reservations.views.reservation_make_qr_ticket import ReservationTicketView
 from reservations.views.organizer_reservation_check_in_ticket import OrganizerReservationCheckInView
